improved Installer

+++ Axon.Installer/Program.cs
@@ -1,4 +1,5 @@
 ﻿using System;
+using System.Diagnostics;
 using System.Net;
 using System.Net.Http;
 using System.Windows.Forms;
@@ -7,12 +8,100 @@ namespace Axon.Installer;
 
 public static class Programm
 {
-    //Just my personal reference :D
-    //D:\Programme\Steam\steamapps\common\SCP Secret Laboratory
     [STAThread]
     public static void Main(string[] args)
     {
-        string gameVersion = "13.4.2";
+        var gameVersion = "";
+        var selectDirectory = false;
+        var installMelonLoader = true;
+        var installAxon = true;
+
+        //Select Game version
+        var client = new HttpClient();
+        if (!Uri.TryCreate("https://raw.githubusercontent.com/AxonSL/SLClientPatches/main/versions.txt", UriKind.Absolute, out var uriResult))
+            throw new InvalidOperationException("URI is invalid.");
+        Console.WriteLine("Getting compatible versions...");
+        var versionsString = client.GetStringAsync(uriResult).GetAwaiter().GetResult();
+        var versions = versionsString.Split('\n');
+        Console.WriteLine("Which game Version do you want to patch?");
+        Console.WriteLine("These versions are available: " + string.Join(",", versions));
+        for(; ; )
+        {
+            gameVersion = Console.ReadLine();
+
+            if (!versions.Any(x => x == gameVersion))
+            {
+                Console.WriteLine("Invalid Version");
+                continue;
+            }
+            break;
+        }
+
+        //Configuration
+        Console.WriteLine($"Current configuration:\nGameVersion: {gameVersion}\nManualDirectory: {selectDirectory}\nInstall MelonLoader: {installMelonLoader}\nInstall Axon: {installAxon}");
+        for(; ; )
+        {
+            Console.WriteLine("Do you want to edit the installation Configuration?[edit/keep]");
+            var input = Console.ReadLine().ToLower();
+
+            if (input == "keep")
+                break;
+
+            if (input != "edit") continue;
+
+            Console.WriteLine("Do you want to select a directory yourself?(yes/no)");
+
+            switch (Console.ReadLine().ToLower())
+            {
+                case "y":
+                case "yes":
+                    selectDirectory = true;
+                    break;
+
+                case "n":
+                case "no":
+                    selectDirectory = false;
+                    break;
+            }
+
+            Console.WriteLine("Do you want to install Melonloader(yes/no)");
+
+            switch (Console.ReadLine().ToLower())
+            {
+                case "y":
+                case "yes":
+                    installMelonLoader = true;
+                    break;
+
+                case "n":
+                case "no":
+                    installMelonLoader = false;
+                    break;
+            }
+
+            if (!installMelonLoader)
+            {
+                installAxon = false;
+                break;
+            }
+
+            Console.WriteLine("Do you want to install Axon(yes/no)");
+
+            switch (Console.ReadLine().ToLower())
+            {
+                case "y":
+                case "yes":
+                    installAxon = true;
+                    break;
+
+                case "n":
+                case "no":
+                    installAxon = false;
+                    break;
+            }
+            break;
+        }
+        Console.WriteLine($"Current configuration:\nGameVersion: {gameVersion}\nManualDirectory: {selectDirectory}\nInstall MelonLoader: {installMelonLoader}\nInstall Axon: {installAxon}");
 
         //Ask for SCPSL.Exe
         var dialog = new OpenFileDialog();
@@ -23,11 +112,22 @@ public static class Programm
         var exePath = dialog.FileName;
         var vanillaPath = Path.GetDirectoryName(exePath);
 
-        //Ask for new install directroy
-        var folderDialog = new FolderBrowserDialog();
-        folderDialog.Description = "Select a Directory in which the modded version will be installed";
-        if (folderDialog.ShowDialog() != DialogResult.OK) return;
-        var moddedPath = folderDialog.SelectedPath;
+        //Install directroy
+        var moddedPath = "";
+        if (selectDirectory)
+        {
+            var folderDialog = new FolderBrowserDialog();
+            folderDialog.Description = "Select a Directory in which the modded version will be installed";
+            if (folderDialog.ShowDialog() != DialogResult.OK) return;
+            moddedPath = folderDialog.SelectedPath;
+        }
+        else
+        {
+            moddedPath = Path.Combine(Directory.GetParent(exePath).Parent.FullName, "SCP Secret Laboratory AxonClient");
+        }
+
+        Console.WriteLine("Client will be installed at:");
+        Console.WriteLine(moddedPath);
 
         //Copy Sl vanilla to the modded Path
         CopyFilesRecursively(new DirectoryInfo(vanillaPath), new DirectoryInfo(moddedPath));
@@ -60,19 +160,26 @@ public static class Programm
         }
 
         //Install Melonloader
-        var zip = Path.Combine(moddedPath, "MelonLoader.x64.zip");
-        if (!File.Exists(zip))
-            File.Create(zip).Close();
-        Console.WriteLine("Downloading Melonloader");
-        DownloadFile("https://github.com/LavaGang/MelonLoader/releases/latest/download/MelonLoader.x64.zip", zip);
-        Console.WriteLine("Extracting Melonloader");
-        System.IO.Compression.ZipFile.ExtractToDirectory(zip, moddedPath);
-        File.Delete(zip);
+        if(installMelonLoader)
+        {
+            var zip = Path.Combine(moddedPath, "MelonLoader.x64.zip");
+            if (!File.Exists(zip))
+                File.Create(zip).Close();
+            Console.WriteLine("Downloading Melonloader");
+            DownloadFile("https://github.com/LavaGang/MelonLoader/releases/latest/download/MelonLoader.x64.zip", zip);
+            Console.WriteLine("Extracting Melonloader");
+            System.IO.Compression.ZipFile.ExtractToDirectory(zip, moddedPath);
+            Console.WriteLine("Deleting zip");
+            File.Delete(zip);
+        }
+
+        //TODO: Install Axon
 
         //Finally
-        Console.WriteLine("You game is now patched. In order to start it open a cmd and type \"SCPSL.exe -noauth\" in it.");
-        Console.WriteLine("Have Fun :D");
+        Console.WriteLine("Your game is now patched." + (installAxon ? "" : " In order to start it open a cmd and type \"SCPSL.exe -noauth\" in it."));
+        Console.WriteLine("We also recommend to add the game to your Steam library");
         Console.WriteLine("Press any Key to Close");
+        Process.Start("explorer.exe", moddedPath);
         Console.ReadKey();
     }
 
@@ -116,14 +223,10 @@ public static class Programm
         var gameassemblyPatches = client.GetStringAsync(uriResult).GetAwaiter().GetResult();
         var patches = gameassemblyPatches.Split("\n");
 
-        Console.WriteLine(patches.Length);
-
         foreach (var patch in patches)
             Console.WriteLine(patch);
 
         var fileBytes = File.ReadAllBytes(file);
-
-        Console.WriteLine("Found " + fileBytes.Length + " bytes");
 
         for(var i = 1; i < patches.Length; i++)
         {
